requirements.txt
streamlit
biopython
watchdog